Fix __ballot_sync cast for HIP/ROCm compatibility
Add explicit static_cast to unsigned long long for __ballot_sync mask
parameter to satisfy HIP runtime requirement that sizeof(MaskT) == 8.
cudf::LANE_MASK_ALL is unsigned int (4 bytes) but __ballot_sync
expects 64-bit mask on AMD GPUs.

+++ cpp/src/rolling/jit/kernel.cu
@@ -15,7 +15,7 @@
  */
 // MIT License
 //
-// Modifications Copyright (C) 2023-2025 Advanced Micro Devices, Inc. All rights reserved.
+// Modifications Copyright (C) 2023-2026 Advanced Micro Devices, Inc. All rights reserved.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -78,7 +78,7 @@ CUDF_KERNEL void gpu_rolling_new(cudf::size_type nrows,
 
   cudf::size_type warp_valid_count{0};
 
-  auto active_threads = __ballot_sync(cudf::LANE_MASK_ALL, i < nrows);
+  auto active_threads = __ballot_sync(static_cast<unsigned long long>(cudf::LANE_MASK_ALL), i < nrows);
   while (i < nrows) {
     int64_t const preceding_window = get_window(preceding_window_begin, i);
     int64_t const following_window = get_window(following_window_begin, i);
